% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{aardl}
\alias{aardl}
\title{Augmented ARDL(AARDL) cointegration test}
\usage{
aardl(model, case = 3, PF = F, Pt = F, signif_level = "auto")
}
\arguments{
\item{model}{The kardl obejct}

\item{case}{Numeric or character. Specifies the case number for the test. Acceptable values are 1, 3, 5, and "auto". If "auto" is chosen, the function determines the case automatically based on the model.}

\item{PF}{The PSS F test object if available}

\item{Pt}{The PSS t test object if available}

\item{signif_level}{Character or numeric. Specifies the significance level to be used in the function.
Acceptable values are "auto", "0.10", "0.1", "0.05", "0.025", and "0.01".
If a numeric value is provided, it will be converted to a character string.
If "auto" is chosen, the function determines the significance level automatically.
Invalid values will result in an error.}
}
\value{
A list containing the results of the AARDL cointegration test, including:
\itemize{
\item \code{type}: The type of test performed, which is "cointegration".
\item \code{case}: The case number used in the test (1, 3, 5, or "auto").
\item \code{statistic}: The F-statistic value calculated from the test.
\item \code{k}: The number of long-run variables in the model.
\item \code{Cont}: The conclusion of the test, indicating whether cointegration is present, inconclusive, or absent.
\item \code{BoundNum}: A numeric representation of the conclusion, where 1 indicates cointegration, 0 indicates inconclusive, and -1 indicates no cointegration.
\item \code{BoundNumAll}: A numeric representation of the conclusion for all tests, where 1 indicates cointegration, 0 indicates inconclusive, and -1 indicates no cointegration.
\item \code{siglvl}: The significance level used in the test, either "auto" or one of the specified numeric levels.
\item \code{siglvls}: A vector of significance levels used in the test for all tests.
\item \code{criticalValues}: A vector of critical values for the test, corresponding to the significance levels.
\item \code{parameter}: The names of the long-run variables in the model.
\item \code{Fmodel}: The linear hypothesis model used in the test.
\item \code{Bound}: The conclusion of the test, indicating whether cointegration is present, inconclusive, or absent.
\item \code{Pt}: The t-statistic value calculated from the PSS t test.
\item \code{PF}: The F-statistic value calculated from the PSS F test.
\item \code{warnings}: Any warnings generated during the test, such as sample size concerns.
\item \code{method}: The method used for the test, which is "AARDL".

}
}
\description{
This function performs the Augmented Autoregressive Distributed Lag (AARDL) bounds test for cointegration, offering several advantages over other methods. First, it allows variables to be integrated at levels I(0), I(1), or a mix of both, without requiring variables to be integrated beyond I(1). Second, it does not necessitate the dependent variable to be I(1). Additionally, it conducts three tests to verify cointegration: a general F-test, a t-test for the dependent variable, and an F-test for independent variables. Finally, the AARDL test remains robust against small sample sizes and endogeneity issues.
}
\examples{


kardl_model<-kardl(imf_example_data,
                    CPI~ER+PPI+asym(ER)+deterministic(covid)+trend,
                    mode=c(1,2,3,0))
A<- aardl(kardl_model)
# critical Values are
A$criticalValues

#' # Using magrittr

library(magrittr)
imf_example_data \%>\% kardl(CPI~ER+PPI+asym(ER)+deterministic(covid)+trend,
                    mode=c(1,2,3,0)) \%>\%  aardl()

 # Getting details of the test.
 mySummary<-summary(A)
 mySummary
# The null hypothesis :
\dontrun{
mySummary$H0
}

#' # Using magrittr
imf_example_data \%>\% kardl(CPI~ER+PPI+asym(ER)+deterministic(covid)+trend,
                    mode=c(1,2,3,0)) \%>\%  aardl() \%>\% summary()
}
\references{
McNown, R., Sam, C. Y., & Goh, S. K. (2018). Bootstrapping the autoregressive distributed lag test for cointegration. Applied Economics, 50(13), 1509-1521.

Sam, C. Y., McNown, R., & Goh, S. K. (2019). An augmented autoregressive distributed lag bounds test for cointegration. Economic Modelling, 80, 130-141.
}
\seealso{
\code{\link{pssf}}  \code{\link{psst}}  \code{\link{banerjee}}     \code{\link{narayan}}
}
