% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kardl-package.R
\docType{package}
\name{kardl-package}
\alias{kardl-package}
\title{kardl: Make Symmetric and Asymmetric ARDL Estimations}
\description{
Implements estimation procedures for Autoregressive Distributed Lag (ARDL) and Nonlinear ARDL (NARDL) models, which allow researchers to investigate both short- and long-run relationships in time series data under mixed orders of integration. The package supports simultaneous modeling of symmetric and asymmetric regressors, flexible treatment of short-run and long-run asymmetries, and automated equation handling. It includes several cointegration testing approaches such as the Pesaran-Shin-Smith F and t bounds tests, and narayan test. Methodological foundations are provided in Pesaran, Shin, and Smith (2001) \doi{10.1016/S0304-4076(01)00049-5} and Shin, Yu, and Greenwood-Nimmo (2014, ISBN:9780123855079).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://karamelikli.github.io/kardl/}
  \item \url{https://github.com/karamelikli/kardl}
  \item Report bugs at \url{https://github.com/karamelikli/kardl/issues}
}

}
\author{
\strong{Maintainer}: Huseyin karamelikli \email{hakperest@gmail.com} (\href{https://orcid.org/0000-0001-7622-0972}{ORCID})

Authors:
\itemize{
  \item Huseyin karamelikli \email{hakperest@gmail.com} (\href{https://orcid.org/0000-0001-7622-0972}{ORCID})
  \item Huseyin Utku Demir (\href{https://orcid.org/0000-0002-9140-0362}{ORCID})
}

}
\keyword{internal}
