% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{kardl_reset}
\alias{kardl_reset}
\title{Reset kardl Package Options to Default Values}
\usage{
kardl_reset(exclude = NULL)
}
\arguments{
\item{exclude}{Character vector of setting names that should not be reset.
These settings retain their current values. By default, all settings are
reset.}
}
\value{
A list of the settings after reset, returned invisibly.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This function resets kardl package options to their default values.
}
\examples{
kardl_set(criterion = "BIC", different_asym_lag = FALSE)

# Reset all settings to defaults except "criterion"
kardl_reset(exclude = "criterion")

# Get the current settings to verify the reset
print(kardl_get("criterion"))

# This will show "BIC" since it was excluded from the reset,
# while other settings will be reset to their defaults.
print(kardl_get("different_asym_lag"))
}
\seealso{
\code{\link{kardl_set}}, \code{\link{kardl_get}}
}
