% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kardl.R
\name{model_criterion}
\alias{model_criterion}
\title{Model Selection Criteria}
\usage{
model_criterion(lm_model, cr, ...)
}
\arguments{
\item{lm_model}{An object containing the fitted model. The object should
include at least:
\itemize{
\item \code{lm_model$model} - the actual fitted model object
(e.g., from \code{lm}, \code{glm}).
\item \code{k} - the number of estimated parameters.
\item \code{n} - the sample size.
}}

\item{cr}{A character string specifying the criterion to compute.
Options are \code{"AIC"}, \code{"BIC"}, \code{"AICc"},\code{"HQ"}
and \code{"AdjR2"}. Alternatively, a user-defined function can be
provided. See details below for more information on using custom
criteria.}

\item{...}{Additional arguments passed to the user-defined criterion
function if \code{cr} is a function.}
}
\value{
A numeric value representing the selected criterion, normalized
by the sample size if one of the predefined options is used.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#superseded}{\figure{lifecycle-superseded.svg}{options: alt='[Superseded]'}}}{\strong{[Superseded]}}

Computes a model selection criterion (AIC, BIC, AICc, HQ, or AdjR2) or
applies a user-defined function to evaluate a statistical model.
}
\details{
This function returns model selection criteria used to compare the quality of
different models.
All criteria are defined such that \strong{lower values indicate better
models} (i.e., the goal is minimization).

If you wish to compare models using a maximization approach (e.g.,
log-likelihood),
you can multiply the result by \code{-1}.

Note: The predefined string options (e.g., \code{"AIC"}) are \strong{not}
the same as the built-in R functions \code{AIC()} or \code{BIC()}.
In particular, the values returned by this function are adjusted by
dividing by the sample size \code{n} (i.e., normalized AIC/BIC), which
makes it more comparable across datasets of different sizes.

The function returns:
\itemize{
\item \strong{"AIC"}:  \eqn{ \frac{2k - 2\ell}{n} } Akaike Information
Criterion divided by \code{n}.
\item \strong{"BIC"}:  \eqn{ \frac{\log(n) \cdot k - 2\ell}{n} } Bayesian
Information Criterion divided by \code{n}.
\item \strong{"AICc"}: \eqn{ \frac{2k(k+1)}{n - k - 1} +
 \frac{2k - 2\ell}{n} }
Corrected Akaike Information Criterion divided by \code{n}.
\item \strong{"HQ"}: \eqn{ \frac{2 \log(\log(n)) \cdot k - 2\ell}{n} }
Hannan-Quinn Criterion divided by \code{n}.
\item \strong{"AdjR2"}: Adjusted R-squared, computed as
\eqn{1 - \frac{(1 - R^2)(n - 1)}{n - k - 1}}. In this case, higher values
indicate better models, but the function still returns the negative of the
adjusted R-squared for consistency with the minimization approach of the
other criteria.
}

where:
\itemize{
\item \eqn{k} is the number of parameters,
\item \eqn{n} is the sample size,
\item \eqn{\ell} is the log-likelihood of the model.
}

If \code{cr} is a function, it is called with the fitted model and any
additional arguments passed through \code{...}.
}
\examples{

# Example usage of model_criterion function with a simple linear model
mylm <- lm(mpg ~ wt + hp, data = mtcars)
model_criterion(mylm, AIC)
model_criterion(mylm, "AIC")

# Example usage of model_criterion function with a kardl model
kardl_model <- kardl(
  DriversKilled ~ PetrolPrice + drivers + asym(PetrolPrice) +
    deterministic(law) + trend,
  Seatbelts,
  mode = c(1, 2, 3, 0)
)

# Using AIC as the kardl package's built-in criterion function which is
# different from the base R AIC function.
model_criterion(kardl_model, "AIC")

# Using the base R AIC function directly on the fitted model object
model_criterion(kardl_model, AIC)
# Using the base R AIC function outside of model_criterion to compute AIC for
#  the fitted model
AIC(kardl_model)

# Using BIC as the criterion for the kardl model which is different from the
# base R BIC function.
model_criterion(kardl_model, "BIC")

# Using a custom criterion function that divides AIC by the sample size
my_cr_fun <- function(mod, ...) {
  AIC(mod) / length(mod$model[[1]])
}
model_criterion(kardl_model, my_cr_fun)

}
\seealso{
\code{\link{kardl}}
}
