% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{plot.kardl}
\alias{plot.kardl}
\title{Plot method for CUSUM and CUSUMQ tests}
\usage{
\method{plot}{kardl}(x, ...)
}
\arguments{
\item{x}{An object of class "kardl" created by \code{\link{cusum}} or \code{\link{cusumq}}.}

\item{...}{Additional arguments passed to \code{plot()}.}
}
\description{
Plot method for CUSUM and CUSUMQ tests
}
