% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_versions_data.R
\docType{data}
\name{imf_example_data}
\alias{imf_example_data}
\title{IMF Example Data (Deprecated)}
\format{
A data frame with 470 rows and 4 variables:
\describe{
\item{ER}{Numeric. Exchange rate of Turkey.}
\item{CPI}{Numeric. CPI of Turkey.}
\item{PPI}{Numeric. PPI of Turkey.}
\item{covid}{Integer. COVID-19 dummy variable.}
}
}
\usage{
imf_example_data
}
\description{
\lifecycle{deprecated}

This dataset is \strong{deprecated} and is provided for testing purposes only
for versions of \pkg{kardl} before 2.0.5. It is retained for backward
compatibility.
}
\examples{
data(imf_example_data)
head(imf_example_data)
}
\keyword{datasets}
