% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{kardl_get}
\alias{kardl_get}
\title{Get kardl Package Options}
\usage{
kardl_get(...)
}
\arguments{
\item{...}{Option names to retrieve. If no arguments are provided, all options will be returned.}
}
\value{
If no arguments are provided, returns all options as a list. If specific option names are provided, returns their values.
}
\description{
This function retrieves the current settings of the kardl package. Users can specify option names to get their values or call the function without arguments to retrieve all current settings.
}
\examples{

# Get all options
kardl_get()
# Get specific options
kardl_get("maxlag", "mode")

# Note: In interactive use, avoid calling kardl_get() directly to prevent cluttering the console.

a<-kardl_get()
a$AsymSuffix

# To reset options and then get a default option:
mydefaults <- kardl_reset()
mydefaults$maxlag

}
\seealso{
\code{\link{kardl_set}}, \code{\link{kardl_reset}}
}
