% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{kardl_get}
\alias{kardl_get}
\title{Function to get settings}
\usage{
kardl_get(...)
}
\arguments{
\item{...}{Names of the options to retrieve. If none provided, all options are returned.}
}
\value{
If no arguments are provided, returns all options as a list. If one option is requested, returns its value directly. If multiple options are requested, returns a list of those options.
}
\description{
This function retrieves options from the kardl package settings.
}
\examples{
# Get specific options
kardl_get("criterion", "differentAsymLag")

# Get all options (Note: In interactive use, avoid calling this directly to prevent cluttering the console)
a<-kardl_get()
a$AsymSuffix

}
\seealso{
\code{\link{kardl_set}}, \code{\link{kardl_reset}}
}
