% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{kardl_reset}
\alias{kardl_reset}
\title{Reset kardl Package Options to Default Values}
\usage{
kardl_reset(value = FALSE)
}
\arguments{
\item{value}{If provided and not `FALSE`, the function will return this value after resetting the settings. If not provided or set to `FALSE`, it will return the current settings.}
}
\value{
If resetting options, returns the provided value (if any) or invisibly
returns the current settings as a list.

If resetting options, returns the provided value (if any) or invisibly
returns the current settings as a list.
}
\description{
This function resets all options in the kardl package to their default values.
}
\examples{

# Set some options
kardl_set(criterion = "BIC", differentAsymLag = TRUE)

# Reset to default options
before <- kardl_get("criterion")
out <-kardl_reset()
after <- kardl_get("criterion")
cat("Before reset:", before, "\n")
cat("After reset:", after, "\n")
}
\seealso{
\code{\link{kardl_set}}, \code{\link{kardl_get}}
}
