% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{kardl_set}
\alias{kardl_set}
\title{Set kardl Package Options}
\usage{
kardl_set(...)
}
\arguments{
\item{...}{Named arguments corresponding to the options to be set. Valid option names include those defined in the kardl package settings.}
}
\value{
If no arguments are provided, returns all options as a list. If named arguments are provided, sets those options and returns the updated list.

If no arguments are provided, returns all options as a list. If named arguments are provided, sets those options and returns the updated list.
}
\description{
This function allows users to set options for the kardl package. Users can specify named arguments to set options or call the function without arguments to retrieve all current settings.
}
\examples{
# Set options
kardl_set(maxlag = 5, mode = "grid")
# Get all options
kardl_get()
# Get specific options
kardl_get("maxlag", "mode")

# Note: In interactive use, avoid calling kardl_get() directly to prevent cluttering the console.

\donttest{
kardl_get()
}

# Set custom coefficient naming conventions

MyFormula<-CPI~ER+PPI+asym(ER)+deterministic(covid)+trend
kardl_set(ShortCoef = "L___{lag}.d.{varName}", formula = MyFormula, data = imf_example_data)

kardl_reset()
kardl_get()


}
\seealso{
\code{\link{kardl_get}}, \code{\link{kardl_reset}}
}
