% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multipliers.R
\name{mplier}
\alias{mplier}
\title{Calculating multipliers of estimated model}
\usage{
mplier(kmodel, horizon = 80, minProb = 0)
}
\arguments{
\item{kmodel}{The model produced by kardl function. This is the model object from which the dynamic multipliers will be calculated. The function expects an object of class \code{kardl_lm}, which is the standard output class for linear models estimated using the kardl package. If the input object does not inherit from this class, the function will throw an error, ensuring that it operates on compatible model objects.
The function extracts necessary information from the model, such as coefficients, lag structure, and variable names, to compute the dynamic multipliers. It calculates the short-run coefficients, Lambda values, and omega values based on the model's parameters and lag structure. The output includes a matrix of dynamic multipliers (mpsi), which can be used for further analysis or visualization.}

\item{horizon}{The horizon over which multipliers will be computed. This parameter defines the time frame for the analysis, allowing users to specify how many periods into the future they want to calculate the multipliers for. For example, setting \code{horizon = 40} will compute the multipliers for 40 periods ahead. The function uses this horizon to determine the number of rows in the output matrix of multipliers and to structure the calculations accordingly.}

\item{minProb}{The minimum p-value threshold for including coefficients in the multipliers calculation. Coefficients with p-values above this threshold will be set to zero in the multipliers calculation. This parameter allows users to control the inclusion of coefficients based on their statistical significance. Setting a threshold can help focus the analysis on more relevant variables, but it may also exclude potentially important effects if set too stringently. The default value is \code{0}, which means that all coefficients will be included regardless of their p-values.}
}
\value{
A list containing the following elements:
\itemize{
 \item \strong{mpsi}: A matrix of dynamic multipliers for each
 variable across the specified horizon. The first column typically represents the time horizon (e.g., 0 to 40), while the subsequent columns contain the multiplier values for each variable.
 \item \strong{omega}: A vector of omega values calculated based on the  model's parameters and lag structure. These values are used in the calculation of the dynamic multipliers and provide insight into the persistence of effects over time.
 \item \strong{lambda}: A matrix of Lambda values that represent the short run coefficients and their transformations based on the model's parameters. These values are essential for understanding the immediate effects of changes in the independent variables on the dependent variable.
 \item \strong{horizon}: The horizon over which the multipliers were computed, as specified in the function's input. This value indicates the time frame for which the dynamic multipliers are relevant and can be used for interpretation and analysis of the results.
 \item \strong{model}: The original model object that was passed to the function. This allows users to reference the model used for the calculations and can be useful for further analysis or validation of the results.
 }
}
\description{
By running \code{m<-mplier(kardl_model)}, the object \code{m$mpsi} contains the final multiplier values, \code{m$omega} holds the omega values, and  \code{m$lambda} returns the Lambda values of the model.
}
\details{
The \code{mplier} function computes dynamic multipliers based on the coefficients and lag structure of a model estimated using the \code{kardl} package. The function extracts necessary
information from the model, such as coefficients, lag structure, and variable names, to compute the dynamic multipliers. It calculates the short-run coefficients, Lambda values, and omega values based on the model's parameters and lag structure. The output includes a matrix of dynamic multipliers (mpsi), which can be used for further analysis or visualization.
The dynamic multipliers provide insight into how changes in the independent variables affect the dependent variable over time, allowing for a deeper understanding of the relationships captured by the model. The function also allows users to set a minimum p-value threshold for including coefficients in the calculation, providing flexibility in focusing on statistically significant effects.
The function is designed to work specifically with models estimated using the \code{kardl} package, and it ensures that the input model is of the correct class before proceeding with the calculations. The resulting multipliers can be used for various applications, including forecasting, policy analysis, and understanding the dynamic relationships between variables in time series data.

\deqn{
{{\psi_{h}^{+} = {\sum_{i = 0}^{h}\frac{\partial y_{t + i}}{\partial x_{t}^{+}}}} ; {\psi_{h}^{-} = {\sum_{i = 0}^{h}\frac{\partial y_{t + i}}{\partial x_{t}^{-}}}}}
}
The above equation represents the cumulative dynamic multipliers for positive and negative changes in an independent variable (x) on the dependent variable (y) over a specified horizon (h). The multipliers are calculated as the sum of the partial derivatives of the dependent variable with respect to the positive and negative changes in the independent variable across the time horizon. These multipliers provide insight into how changes in the independent variable affect the dependent variable over time, allowing for a deeper understanding of the relationships captured by the model.
}
\examples{

# This example demonstrates how to use the mplier function to calculate dynamic multipliers
# from a model estimated using the kardl package. The example includes fitting a model with
# the kardl function, calculating the multipliers, and visualizing the results using both
# base R plotting and ggplot2.

 kardl_model<-kardl(imf_example_data, CPI~ER+PPI+asym(ER)+deterministic(covid)+trend,
 mode=c(1,2,3,0))
 m<-mplier(kardl_model,40)
 head(m$mpsi)
 head(m$omega)
 head(m$lambda)

 # Displaying the summary of the multipliers object
 summary(m)

 # Visualize the dynamic multipliers

plot(m)

 # For plotting specific variables, you can specify them in the plot function. For example,
 # to plot the multipliers for the variable "ER":

 plot(m, variable = "ER")


}
\seealso{
\code{\link{bootstrap}}
}
